LicenseValidationService
