+++ layers/GatoGraphQLForWP/plugins/gatographql/config/services.yaml
@@ -40,6 +40,9 @@ services:
     GatoGraphQL\GatoGraphQL\Log\LoggerInterface:
         class: \GatoGraphQL\GatoGraphQL\Log\Logger
 
+    GatoGraphQL\GatoGraphQL\Marketplace\LicenseValidationServiceInterface:
+        class: \GatoGraphQL\GatoGraphQL\Marketplace\LicenseValidationService
+
     GatoGraphQL\GatoGraphQL\Services\:
         resource: ../src/Services/*
